# 04-laravel-master/app/Models/Comment.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Comment extends Model
{
    use HasFactory;

    public $timestamps = false;

    protected $fillable = [
        "userId",
        "recipeId",
        "commentContent"
    ];

    protected $primaryKey = "commentId";
}